api_gateway/requirements.txt
flask==2.0.1
werkzeug==2.0.3
requests==2.26.0
pyjwt==2.1.0